Add files via upload
Currently, China is in the stage of rapid urban development, and construction activities are increasing. This case is associated with a large amount of construction waste, which causes many problems such as occupying valuable land resources, creating air pollution and consuming raw materials in its management. In this study, we consider Guangzhou as an example. Basically, system dynamics was used to create an environmental assessment model for construction waste, and Vensim software was used to simulate and analyze the environmental, economic, and social effects of different construction waste disposal methods.